requirements.txt
﻿﻿﻿﻿﻿# Headout Scraper Dependencies
playwright==1.40.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.67.1
openpyxl==3.1.5
pytest==8.3.3
